+++ packages/es-client/search.js
@@ -140,8 +140,6 @@ class BaseSearch {
 
     if (this.type === process.env.CollectionsTable) {
       this.hash = 'collectionName';
-    } else if (this.type === process.env.PdrsTable) {
-      this.hash = 'pdrName';
     }
   }