+++ tf-modules/archive/api.tf
@@ -5,7 +5,6 @@ resource "aws_ssm_parameter" "dynamo_table_names" {
     AccessTokensTable          = var.dynamo_tables.access_tokens.name
     AsyncOperationsTable       = var.dynamo_tables.async_operations.name
     CollectionsTable           = var.dynamo_tables.collections.name
-    PdrsTable                  = var.dynamo_tables.pdrs.name
     ProvidersTable             = var.dynamo_tables.providers.name
     ReconciliationReportsTable = var.dynamo_tables.reconciliation_reports.name
     RulesTable                 = var.dynamo_tables.rules.name
@@ -20,7 +19,6 @@ locals {
     AccessTokensTable          = var.dynamo_tables.access_tokens.name
     AsyncOperationsTable       = var.dynamo_tables.async_operations.name
     CollectionsTable           = var.dynamo_tables.collections.name
-    PdrsTable                  = var.dynamo_tables.pdrs.name
     ProvidersTable             = var.dynamo_tables.providers.name
     ReconciliationReportsTable = var.dynamo_tables.reconciliation_reports.name
     RulesTable                 = var.dynamo_tables.rules.name