+++ tf-modules/archive/db_indexer.tf
@@ -24,7 +24,6 @@ resource "aws_lambda_function" "db_indexer" {
       CMR_HOST                   = var.cmr_custom_host
       CollectionsTable           = var.dynamo_tables.collections.name
       AsyncOperationsTable       = var.dynamo_tables.async_operations.name
-      PdrsTable                  = var.dynamo_tables.pdrs.name
       ProvidersTable             = var.dynamo_tables.providers.name
       ReconciliationReportsTable = var.dynamo_tables.reconciliation_reports.name
       RulesTable                 = var.dynamo_tables.rules.name