+++ tf-modules/data-persistence/outputs.tf
@@ -12,10 +12,6 @@ output "dynamo_tables" {
       name = aws_dynamodb_table.collections_table.name
       arn  = aws_dynamodb_table.collections_table.arn
     }
-    pdrs = {
-      name = aws_dynamodb_table.pdrs_table.name
-      arn  = aws_dynamodb_table.pdrs_table.arn
-    }
     providers = {
       name = aws_dynamodb_table.providers_table.name
       arn  = aws_dynamodb_table.providers_table.arn